refactor(frontend): extract shared filter multi-select component
